+++ crate/params/src/params_spec.rs
@@ -4,8 +4,8 @@ use peace_resources::{resources::ts::SetUp, BorrowFail, Resources};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
 use crate::{
-    FieldWiseSpecRt, MappingFn, MappingFnImpl, Params, ParamsResolveError, ValueResolutionCtx,
-    ValueSpecRt,
+    AnySpecDataType, AnySpecRt, FieldWiseSpecRt, MappingFn, MappingFnImpl, Params,
+    ParamsResolveError, ValueResolutionCtx, ValueSpecRt,
 };
 
 /// How to populate a field's value in an item spec's params.
@@ -115,7 +115,7 @@ where
             Self::Stored => f.write_str("Stored"),
             Self::Value { value } => f.debug_tuple("Value").field(value).finish(),
             Self::InMemory => f.write_str("From"),
-            Self::MappingFn(_) => f.debug_tuple("MappingFn").field(&"..").finish(),
+            Self::MappingFn(mapping_fn) => f.debug_tuple("MappingFn").field(mapping_fn).finish(),
             Self::FieldWise { field_wise_spec } => {
                 f.debug_tuple("FieldWise").field(field_wise_spec).finish()
             }
@@ -189,6 +189,59 @@ where
                 .map(|t| t.map(T::Partial::from).unwrap_or_else(T::Partial::default)),
             ParamsSpec::FieldWise { field_wise_spec } => {
                 field_wise_spec.resolve_partial(resources, value_resolution_ctx)
+            }
+        }
+    }
+}
+
+impl<T> AnySpecRt for ParamsSpec<T>
+where
+    T: Params<Spec = ParamsSpec<T>>
+        + Clone
+        + Debug
+        + Serialize
+        + DeserializeOwned
+        + Send
+        + Sync
+        + 'static,
+{
+    fn is_usable(&self) -> bool {
+        match self {
+            Self::Stored => false,
+            Self::Value { .. } | Self::InMemory => true,
+            Self::MappingFn(mapping_fn) => mapping_fn.is_valued(),
+            Self::FieldWise { field_wise_spec } => field_wise_spec.is_usable(),
+        }
+    }
+
+    fn merge(&mut self, other_boxed: &dyn AnySpecDataType)
+    where
+        Self: Sized,
+    {
+        let other: Option<&Self> = other_boxed.downcast_ref();
+        let Some(other) = other else {
+            let self_ty_name = tynm::type_name::<Self>();
+            panic!("Failed to downcast value into `{self_ty_name}`. Value: `{other_boxed:#?}`.");
+        };
+        match self {
+            // Use the spec that was previously stored
+            // (as opposed to previous value).
+            Self::Stored => *self = other.clone(),
+
+            // Use set value / no change on these variants
+            Self::Value { .. } | Self::InMemory | Self::MappingFn(_) => {}
+
+            //
+            Self::FieldWise { field_wise_spec } => {
+                match other {
+                    // Don't merge stored field wise specs over provided specs.
+                    Self::Stored | Self::Value { .. } | Self::InMemory | Self::MappingFn(_) => {}
+
+                    // Merge specs fieldwise.
+                    Self::FieldWise {
+                        field_wise_spec: field_wise_spec_other,
+                    } => AnySpecRt::merge(field_wise_spec, field_wise_spec_other),
+                }
             }
         }
     }