+++ crate/params/src/params_spec_fieldless.rs
@@ -4,7 +4,8 @@ use peace_resources::{resources::ts::SetUp, BorrowFail, Resources};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    MappingFn, MappingFnImpl, ParamsFieldless, ParamsResolveError, ValueResolutionCtx, ValueSpecRt,
+    AnySpecDataType, AnySpecRt, MappingFn, MappingFnImpl, ParamsFieldless, ParamsResolveError,
+    ValueResolutionCtx, ValueSpecRt,
 };
 
 /// How to populate a field's value in an item spec's params.
@@ -95,7 +96,7 @@ where
             Self::Stored => f.write_str("Stored"),
             Self::Value { value } => f.debug_tuple("Value").field(value).finish(),
             Self::InMemory => f.write_str("From"),
-            Self::MappingFn(_) => f.debug_tuple("MappingFn").field(&"..").finish(),
+            Self::MappingFn(mapping_fn) => f.debug_tuple("MappingFn").field(mapping_fn).finish(),
         }
     }
 }
@@ -171,9 +172,53 @@ where
     }
 }
 
+impl<T> AnySpecRt for ParamsSpecFieldless<T>
+where
+    T: ParamsFieldless<Spec = ParamsSpecFieldless<T>>
+        + Clone
+        + Debug
+        + Serialize
+        + Send
+        + Sync
+        + 'static,
+{
+    fn is_usable(&self) -> bool {
+        match self {
+            Self::Stored => false,
+            Self::Value { .. } | Self::InMemory => true,
+            Self::MappingFn(mapping_fn) => mapping_fn.is_valued(),
+        }
+    }
+
+    fn merge(&mut self, other_boxed: &dyn AnySpecDataType)
+    where
+        Self: Sized,
+    {
+        let other: Option<&Self> = other_boxed.downcast_ref();
+        let Some(other) = other else {
+            let self_ty_name = tynm::type_name::<Self>();
+            panic!("Failed to downcast value into `{self_ty_name}`. Value: `{other_boxed:#?}`.");
+        };
+        match self {
+            // Use the spec that was previously stored
+            // (as opposed to previous value).
+            Self::Stored => *self = other.clone(),
+
+            // Use set value / no change on these variants
+            Self::Value { .. } | Self::InMemory | Self::MappingFn(_) => {}
+        }
+    }
+}
+
 impl<T> ValueSpecRt for ParamsSpecFieldless<T>
 where
-    T: ParamsFieldless<Spec = ParamsSpecFieldless<T>> + Clone + Debug + Send + Sync + 'static,
+    T: ParamsFieldless<Spec = ParamsSpecFieldless<T>>
+        + Clone
+        + Debug
+        + Serialize
+        + Send
+        + Sync
+        + 'static,
     T::Partial: From<T>,
     T: TryFrom<T::Partial>,
 {