+++ crate/rt_model_core/src/error.rs
@@ -111,6 +111,7 @@ pub enum Error {
                 item_spec_ids_with_no_params_specs,
                 params_specs_provided_mismatches,
                 params_specs_stored_mismatches.as_ref(),
+                params_specs_not_usable,
             ))
         )
     )]
@@ -121,6 +122,10 @@ pub enum Error {
         params_specs_provided_mismatches: ParamsSpecs,
         /// Stored params specs with no matching item spec ID in the flow.
         params_specs_stored_mismatches: Option<ParamsSpecs>,
+        /// Item spec IDs which had a mapping function previously provided in
+        /// its params spec, but on a subsequent invocation nothing was
+        /// provided.
+        params_specs_not_usable: Vec<ItemSpecId>,
     },
 
     /// In a `MultiProfileSingleFlow` diff, neither profile had `Params::Specs`
@@ -566,6 +571,7 @@ fn params_specs_mismatch_display(
     item_spec_ids_with_no_params: &[ItemSpecId],
     params_specs_provided_mismatches: &ParamsSpecs,
     params_specs_stored_mismatches: Option<&ParamsSpecs>,
+    params_specs_not_usable: &[ItemSpecId],
 ) -> String {
     let mut items = Vec::<String>::new();
 
@@ -608,6 +614,21 @@ fn params_specs_mismatch_display(
                 {params_specs_stored_mismatches_list}\n",
             ));
         }
+    }
+
+    if !params_specs_not_usable.is_empty() {
+        items.push(format!(
+            "The following item specs either have not had a params spec provided previously,\n\
+            or previously had a mapping function provided which cannot be automatically loaded.\n\
+            So the current execution requires the spec to be provided:\n\
+            \n\
+            {}\n",
+            params_specs_not_usable
+                .iter()
+                .map(|item_spec_id| format!("* {item_spec_id}"))
+                .collect::<Vec<String>>()
+                .join("\n")
+        ));
     }
 
     items.join("\n")